Bugfixes: - cannot log `entity.name` as the parameter `entity` might be a string, too ... :-( - reverted back last change ... (?)

+++ lua/spawning.lua
@@ -49,7 +49,7 @@ local function spawn_entity(entity, relative_position, center, surface, extra_op
   end
 
   local entity_name = expressions.entity(entity, vars)
-  if debug_log then log(string.format("[spawn_entity]: entity_name='%s',entity.name='%s'", entity_name, entity.name)) end
+  if debug_log then log(string.format("[spawn_entity]: entity_name='%s'", entity_name)) end
 
   if _G["prototypes"].entity[entity_name] == nil then
     log(string.format("[spawn_entity]: entity_name='%s' does not exist!", entity_name))
@@ -256,11 +256,11 @@ local function clear_area(half_size, center, surface)
 
   for _, entity in pairs(surface.find_entities_filtered({area = area, type = {"resource"}, invert = true})) do
     if debug_log and not entity.valid then log("[clear_area]: Found an invalid entity ...") end
-    if debug_log and entity.valid then log(string.format("[clear_area]: Found entity.type='%s',entity.name='%s' ...", entity.type, entity.name)) end
+    elseif debug_log and entity.valid then log(string.format("[clear_area]: Found entity.type='%s',entity.name='%s' ...", entity.type, entity.name)) end
 
-    if entity.type ~= "tree" or math.random() < (half_size / 14) then
+    if (entity.valid and entity.type ~= "tree") or math.random() < (half_size / 14) then
       if debug_log and not entity.valid then log("[clear_area]: Destroying invalid  ...") end
-      if debug_log and entity.valid then log(string.format("[clear_area]: Destroying entity.name='%s' ...", entity.name)) end
+      elseif debug_log and entity.valid then log(string.format("[clear_area]: Destroying entity.name='%s' ...", entity.name)) end
 
       entity.destroy({do_cliff_correction = true, raise_destroy = true})
     end
